+++ src/spaces/grid_single.jl
@@ -27,7 +27,10 @@ with non-zero IDs, either positive or negative. This is not checked internally.
 
 All arguments and keywords behave exactly as in [`GridSpace`](@ref).
 """
-function GridSpaceSingle(d::NTuple{D,Int}; periodic = true, metric = :chebyshev) where {D}
+function GridSpaceSingle(d::NTuple{D,Int};
+        periodic::Union{Bool,NTuple{D,Bool}} = true,
+        metric = :chebyshev
+    ) where {D}
     s = zeros(Int, d)
     return GridSpaceSingle{D,periodic}(s, metric,
         Dict{Int,Vector{NTuple{D,Int}}}(),
@@ -113,6 +116,29 @@ function nearby_ids(pos::NTuple{D, Int}, model::ABM{<:GridSpaceSingle{D,false}},
     else
         ids_iterator = (stored_ids[p...] for p in position_iterator 
                         if checkbounds(Bool, stored_ids, p...) && stored_ids[p...] != 0)
+    end
+    return ids_iterator
+end
+
+function nearby_ids(pos::NTuple{D, Int}, model::ABM{<:GridSpaceSingle{D,P}}, r = 1,
+        get_offset_indices = offsets_within_radius # internal, see last function
+    ) where {D,P}
+    nindices = get_offset_indices(model, r)
+    stored_ids = abmspace(model).stored_ids
+    space_size = size(stored_ids)
+    position_iterator = (pos .+ β for β in nindices)
+    # check if we are far from the wall to skip bounds checks
+    if all(i -> r < pos[i] <= space_size[i] - r, 1:D)
+        ids_iterator = (stored_ids[p...] for p in position_iterator
+                        if stored_ids[p...] != 0)
+    else
+        ids_iterator = (
+            checkbounds(Bool, stored_ids, p...) ?
+            stored_ids[p...] : stored_ids[mod1.(p, space_size)...]
+            for p in position_iterator
+            if stored_ids[mod1.(p, space_size)...] != 0 &&
+            all(P[i] || checkbounds(Bool, axes(stored_ids, i), p[i]) for i in 1:D)
+        )
     end
     return ids_iterator
 end